Worked around the most annoying warning.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@48552 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/CsConvert.h
@@ -70,7 +70,7 @@ public:
 				 bool ignoreTrailingSpaces = false)
 	{
 		dst.getBuffer(convertLength(srcLen));
-		dst.resize(convert(srcLen, src, dst.getCapacity(), dst.begin(), badInputPos, ignoreTrailingSpaces));
+		dst.resize(convert(srcLen, src, (ULONG) dst.getCapacity(), dst.begin(), badInputPos, ignoreTrailingSpaces));
 	}
 
 	// CVC: Beware of this can of worms: csconvert_convert gets assigned
